LayoutTests/platform/gtk/fast/canvas/fillText-shadow-expected.txt

Tests that (vertical) shadow offsets are applied correctly when using fillText() regardless of blur amount.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing with no transform
Testing with blur of 0 pixels
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
Testing with blur of 1 pixels
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
Testing with blur of 5 pixels
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
Testing with scale transform
Testing with blur of 0 pixels
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
FAIL imgdata[3] should be 255. Was 191.
Testing with blur of 1 pixels
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
Testing with blur of 5 pixels
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
PASS successfullyParsed is true

TEST COMPLETE
